Delete bogus tests that use yield.
Putting a yield into a function definition turns the function into a generator
function. Calling a generator function returns a generator, but doesn't
execute any code in the generator function: only when the generator is iterated
is the generator function executed.

So when unittest calls a test method that uses yield it gets back a generator.
It doesn't know a thing about generators. So it ignores it and moves on. The
test passes because it didn't raise an exception.

This means that a unittest test method that is a generator function is not
executed at all and is useless.

Example:

class TestBogus(unittest.TestCase):

    def test_bogus(self):
        yield
        assert False

The above test would pass.