Fix bug related to default parameter value for Master constructor
In Python default parameter values are always evaluated when, and only
when, the def statement they belong to is executed.

From the official documentation:

"Default parameter values are evaluated when the function definition is
executed. This means that the expression is evaluated once, when the
function is defined, and that the same .pre-computed. value is used for
each call. This is especially important to understand when a default
parameter is a mutable object, such as a list or a dictionary: if the
function modifies the object (e.g. by appending an item to a list), the
default value is in effect modified. This is generally not what was
intended. A way around this is to use None as the default, and
explicitly test for it in the body of the function [...]"

http://docs.python.org/3.3/reference/compound_stmts.html#function-definitions